requirements.txt
python-telegram-bot
requests
googletrans==4.0.0-rc1
apscheduler
pytz